+++ python_modules/dagster/dagster_tests/core_tests/runtime_types_tests/config_schema_tests/test_builtin_schemas.py
@@ -18,9 +18,9 @@ from dagster import (
 from dagster.utils.test import get_temp_file_name
 
 
-def _execute_pipeline_with_subset(pipeline, environment_dict, solid_subset):
+def _execute_pipeline_with_subset(pipeline, environment_dict, solid_selection):
     return execute_pipeline(
-        pipeline.get_pipeline_subset_def(solid_subset), environment_dict=environment_dict
+        pipeline.get_pipeline_subset_def(solid_selection), environment_dict=environment_dict
     )
 
 
@@ -100,7 +100,7 @@ def test_int_input_schema_value():
     result = _execute_pipeline_with_subset(
         define_test_all_scalars_pipeline(),
         environment_dict={'solids': {'take_int': {'inputs': {'num': {'value': 2}}}}},
-        solid_subset={'take_int'},
+        solid_selection={'take_int'},
     )
 
     assert result.success
@@ -111,7 +111,7 @@ def test_int_input_schema_raw_value():
     result = _execute_pipeline_with_subset(
         define_test_all_scalars_pipeline(),
         environment_dict={'solids': {'take_int': {'inputs': {'num': 2}}}},
-        solid_subset={'take_int'},
+        solid_selection={'take_int'},
     )
 
     assert result.success
@@ -123,7 +123,7 @@ def test_int_input_schema_failure_wrong_value_type():
         _execute_pipeline_with_subset(
             define_test_all_scalars_pipeline(),
             environment_dict=single_input_env('take_int', 'num', {'value': 'dkjdfkdj'}),
-            solid_subset={'take_int'},
+            solid_selection={'take_int'},
         )
     assert 'Error 1: Invalid scalar at path root:solids:take_int:inputs:num:value' in str(
         exc_info.value
@@ -135,7 +135,7 @@ def test_int_input_schema_failure_wrong_key():
         _execute_pipeline_with_subset(
             define_test_all_scalars_pipeline(),
             environment_dict=single_input_env('take_int', 'num', {'wrong_key': 'dkjdfkdj'}),
-            solid_subset={'take_int'},
+            solid_selection={'take_int'},
         )
     assert 'Error 1: Undefined field "wrong_key" at path root:solids:take_int:inputs:num' in str(
         exc_info.value
@@ -147,7 +147,7 @@ def test_int_input_schema_failure_raw_string():
         _execute_pipeline_with_subset(
             define_test_all_scalars_pipeline(),
             environment_dict=single_input_env('take_int', 'num', 'dkjdfkdj'),
-            solid_subset={'take_int'},
+            solid_selection={'take_int'},
         )
     assert 'Error 1: Invalid scalar at path root:solids:take_int:inputs:num' in str(exc_info.value)
 
@@ -161,7 +161,7 @@ def test_int_json_schema_roundtrip():
         mat_result = _execute_pipeline_with_subset(
             define_test_all_scalars_pipeline(),
             environment_dict=single_output_env('produce_int', {'json': {'path': tmp_file}}),
-            solid_subset={'produce_int'},
+            solid_selection={'produce_int'},
         )
 
         assert mat_result.success
@@ -169,7 +169,7 @@ def test_int_json_schema_roundtrip():
         source_result = _execute_pipeline_with_subset(
             define_test_all_scalars_pipeline(),
             environment_dict=single_input_env('take_int', 'num', {'json': {'path': tmp_file}}),
-            solid_subset={'take_int'},
+            solid_selection={'take_int'},
         )
 
         assert source_result.result_for_solid('take_int').output_value() == 2
@@ -180,7 +180,7 @@ def test_int_pickle_schema_roundtrip():
         mat_result = _execute_pipeline_with_subset(
             define_test_all_scalars_pipeline(),
             environment_dict=single_output_env('produce_int', {'pickle': {'path': tmp_file}}),
-            solid_subset={'produce_int'},
+            solid_selection={'produce_int'},
         )
 
         assert mat_result.success
@@ -188,7 +188,7 @@ def test_int_pickle_schema_roundtrip():
         source_result = _execute_pipeline_with_subset(
             define_test_all_scalars_pipeline(),
             environment_dict=single_input_env('take_int', 'num', {'pickle': {'path': tmp_file}}),
-            solid_subset={'take_int'},
+            solid_selection={'take_int'},
         )
 
         assert source_result.result_for_solid('take_int').output_value() == 2
@@ -198,7 +198,7 @@ def test_string_input_schema_value():
     result = _execute_pipeline_with_subset(
         define_test_all_scalars_pipeline(),
         environment_dict=single_input_env('take_string', 'string', {'value': 'dkjkfd'}),
-        solid_subset={'take_string'},
+        solid_selection={'take_string'},
     )
 
     assert result.success
@@ -210,7 +210,7 @@ def test_string_input_schema_failure():
         _execute_pipeline_with_subset(
             define_test_all_scalars_pipeline(),
             environment_dict=single_input_env('take_string', 'string', {'value': 3343}),
-            solid_subset={'take_string'},
+            solid_selection={'take_string'},
         )
 
     assert 'Invalid scalar at path root:solids:take_string:inputs:string:value' in str(
@@ -222,7 +222,7 @@ def test_float_input_schema_value():
     result = _execute_pipeline_with_subset(
         define_test_all_scalars_pipeline(),
         environment_dict=single_input_env('take_float', 'float_number', {'value': 3.34}),
-        solid_subset={'take_float'},
+        solid_selection={'take_float'},
     )
 
     assert result.success
@@ -234,7 +234,7 @@ def test_float_input_schema_failure():
         _execute_pipeline_with_subset(
             define_test_all_scalars_pipeline(),
             environment_dict=single_input_env('take_float', 'float_number', {'value': '3343'}),
-            solid_subset={'take_float'},
+            solid_selection={'take_float'},
         )
 
     assert 'Invalid scalar at path root:solids:take_float:inputs:float_number:value' in str(
@@ -246,7 +246,7 @@ def test_bool_input_schema_value():
     result = _execute_pipeline_with_subset(
         define_test_all_scalars_pipeline(),
         environment_dict=single_input_env('take_bool', 'bool_value', {'value': True}),
-        solid_subset={'take_bool'},
+        solid_selection={'take_bool'},
     )
 
     assert result.success
@@ -258,7 +258,7 @@ def test_bool_input_schema_failure():
         _execute_pipeline_with_subset(
             define_test_all_scalars_pipeline(),
             environment_dict=single_input_env('take_bool', 'bool_value', {'value': '3343'}),
-            solid_subset={'take_bool'},
+            solid_selection={'take_bool'},
         )
 
     assert 'Invalid scalar at path root:solids:take_bool:inputs:bool_value:value' in str(
@@ -270,7 +270,7 @@ def test_any_input_schema_value():
     result = _execute_pipeline_with_subset(
         define_test_all_scalars_pipeline(),
         environment_dict=single_input_env('take_any', 'any_value', {'value': 'ff'}),
-        solid_subset={'take_any'},
+        solid_selection={'take_any'},
     )
 
     assert result.success
@@ -279,7 +279,7 @@ def test_any_input_schema_value():
     result = _execute_pipeline_with_subset(
         define_test_all_scalars_pipeline(),
         environment_dict=single_input_env('take_any', 'any_value', {'value': 3843}),
-        solid_subset={'take_any'},
+        solid_selection={'take_any'},
     )
 
     assert result.success
@@ -291,7 +291,7 @@ def test_none_string_input_schema_failure():
         _execute_pipeline_with_subset(
             define_test_all_scalars_pipeline(),
             environment_dict=single_input_env('take_string', 'string', None),
-            solid_subset={'take_string'},
+            solid_selection={'take_string'},
         )
 
     assert len(exc_info.value.errors) == 1
@@ -308,7 +308,7 @@ def test_value_none_string_input_schema_failure():
         _execute_pipeline_with_subset(
             define_test_all_scalars_pipeline(),
             environment_dict=single_input_env('take_string', 'string', {'value': None}),
-            solid_subset={'take_string'},
+            solid_selection={'take_string'},
         )
 
     assert 'Value at path root:solids:take_string:inputs:string:value must be not be None' in str(
@@ -321,7 +321,7 @@ def test_string_json_schema_roundtrip():
         mat_result = _execute_pipeline_with_subset(
             define_test_all_scalars_pipeline(),
             environment_dict=single_output_env('produce_string', {'json': {'path': tmp_file}}),
-            solid_subset={'produce_string'},
+            solid_selection={'produce_string'},
         )
 
         assert mat_result.success
@@ -331,7 +331,7 @@ def test_string_json_schema_roundtrip():
             environment_dict=single_input_env(
                 'take_string', 'string', {'json': {'path': tmp_file}}
             ),
-            solid_subset={'take_string'},
+            solid_selection={'take_string'},
         )
 
         assert source_result.result_for_solid('take_string').output_value() == 'foo'
@@ -342,7 +342,7 @@ def test_string_pickle_schema_roundtrip():
         mat_result = _execute_pipeline_with_subset(
             define_test_all_scalars_pipeline(),
             environment_dict=single_output_env('produce_string', {'pickle': {'path': tmp_file}}),
-            solid_subset={'produce_string'},
+            solid_selection={'produce_string'},
         )
 
         assert mat_result.success
@@ -352,7 +352,7 @@ def test_string_pickle_schema_roundtrip():
             environment_dict=single_input_env(
                 'take_string', 'string', {'pickle': {'path': tmp_file}}
             ),
-            solid_subset={'take_string'},
+            solid_selection={'take_string'},
         )
 
         assert source_result.result_for_solid('take_string').output_value() == 'foo'
@@ -362,7 +362,7 @@ def test_string_list_input():
     result = _execute_pipeline_with_subset(
         define_test_all_scalars_pipeline(),
         environment_dict=single_input_env('take_string_list', 'string_list', [{'value': 'foobar'}]),
-        solid_subset={'take_string_list'},
+        solid_selection={'take_string_list'},
     )
 
     assert result.success
@@ -376,7 +376,7 @@ def test_nullable_string_input_with_value():
         environment_dict=single_input_env(
             'take_nullable_string', 'nullable_string', {'value': 'foobar'}
         ),
-        solid_subset={'take_nullable_string'},
+        solid_selection={'take_nullable_string'},
     )
 
     assert result.success
@@ -393,7 +393,7 @@ def test_nullable_string_input_with_none_value():
             environment_dict=single_input_env(
                 'take_nullable_string', 'nullable_string', {'value': None}
             ),
-            solid_subset={'take_nullable_string'},
+            solid_selection={'take_nullable_string'},
         )
 
     assert (
@@ -405,7 +405,7 @@ def test_nullable_string_input_without_value():
     result = _execute_pipeline_with_subset(
         define_test_all_scalars_pipeline(),
         environment_dict=single_input_env('take_nullable_string', 'nullable_string', None),
-        solid_subset={'take_nullable_string'},
+        solid_selection={'take_nullable_string'},
     )
 
     assert result.success